Move some external access types to the UnitTesting.API namespace so that they can be accessed externally via RestrictedInternalsVisibleTo.
