+++ src/Features/Core/Portable/ExternalAccess/UnitTesting/API/UnitTestingInvocationReasons.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Runtime.Serialization;
 
-namespace Microsoft.CodeAnalysis.ExternalAccess.UnitTesting.SolutionCrawler
+namespace Microsoft.CodeAnalysis.ExternalAccess.UnitTesting.Api
 {
     [DataContract]
     internal partial struct UnitTestingInvocationReasons : IEnumerable<string>
